requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
motor==2.5.1
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pymongo==3.12.0
mongoengine==0.23.1
python-dotenv==0.19.0
